# eth/filters: fix failing benchmark-test